refactor: Move XML list wrappers to a separate module.
