+++ src/serde/json/mod.rs
@@ -1,10 +1,10 @@
-pub(crate) mod json_layout;
-pub(crate) mod json_layout_container;
-pub(crate) mod json_layout_variable;
-pub(crate) mod json_model;
-pub(crate) mod json_network;
-pub(crate) mod json_relationship;
-pub(crate) mod json_variable;
+mod json_layout;
+mod json_layout_container;
+mod json_layout_variable;
+mod json_model;
+mod json_network;
+mod json_relationship;
+mod json_variable;
 
 pub(crate) use json_layout::JsonLayout;
 pub(crate) use json_layout_container::JsonLayoutContainer;
