+++ src/serde/xml/mod.rs
@@ -1,9 +1,13 @@
-pub(crate) mod xml_container;
-pub(crate) mod xml_layout;
-pub(crate) mod xml_relationship;
-pub(crate) mod xml_variable;
+mod xml_container;
+mod xml_layout;
+mod xml_lists;
+mod xml_relationship;
+mod xml_variable;
 
 pub(crate) use xml_container::XmlContainer;
 pub(crate) use xml_layout::XmlLayout;
+pub(crate) use xml_lists::XmlContainers;
+pub(crate) use xml_lists::XmlRelationships;
+pub(crate) use xml_lists::XmlVariables;
 pub(crate) use xml_relationship::XmlRelationship;
 pub(crate) use xml_variable::XmlVariable;
